+++ lib/src/event.dart
@@ -1,5 +1,0 @@
-abstract class FlutterFeathersjsEvents {}
-
-class Connected extends FlutterFeathersjsEvents {}
-
-class Disconnected extends FlutterFeathersjsEvents {}