+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## [0.0.1-dev]
+# 0.0.1-dev
 
-- Beta version, don't use it. It still under dev
+- Initial version
+- Not for production use. It still under dev
+
+## 0.0.2-dev
+
+    - Add auth and reAuth for both rest and socketio client
+    - Add socketio methods
+    - Test socketio emit methods
+    - Add global auth and reAuth which don't matter whatever you're calling via socketio or rest